Prevent findOnShelfModal reuqests om material page
by only requesting when the modal is open
